#23 - refactor folder structure layout
