+++ src/model/VAASearchCriteria.ts
